third_party/WebKit/LayoutTests/fast/dom/HTMLDialogElement/dialog-close-event-expected.txt

Test that dialog receives a close event upon closing.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS self is dialog
PASS dialog.open is false
PASS event.cancelable is false
PASS self is dialog
PASS dialog.open is false
PASS event.cancelable is false
PASS dialog.open is false
PASS closedCount is 2
